📝 : Change PieceMovement class to functions...
